Refactored type 3 font, fixed type 3 width checks.
